Match parsed command once
I was matching on the parsed command twice, once in `from_args` and
again in `run`. This was redundant and using the non-explicit arm `_` in
both places.

This change modifies `from_args` to not return an `Err` on an
`Command::Unsupported` and instead to just return the configuration with
that value embedded. Now, `run` handles `Unsupported` commands.

This change also narrows down errors returned from the `run` function to
a `String` value. In the future this should likely be a custom error
type to the application that can contain error chains for the myriad of
problems the application will encounter.