Add ability to "do" todos
Adds the ability to do a todo. Implements doing so by reading the todos
file into memory and iterating each line to see whether or not to write
it back to the todos file or to write it to the done file. A followup
might be to use a buffer writer for the done file.

The format for done todos in the done file is "action datetime_utc
todo":

```txt
x 2022-01-13T13:45:01.390986Z todo text
```

The common action of completing a todo being represented by an "x". More
finalizing actions can be added in the future.