Fix bug in parse_remaining_args
parse_remaining_args was bugged because, without fully reading the
documentation for `reduce`, I didn't realize the function acted like
fold with an initial value of the first item in the iteration. Therefore
the spacing wasn't getting applied correctly and the first two words
were not separated in the resulting todo value.

While we are here, let's rename `target` to `todo` since that's what it
is.